cambio de ruta de listar detalle de entrada y cambio de la migracion detalle de entrada a centavos pero no funciona del todo correcto

+++ database/migrations/2025_03_18_013446_create_detalle_entradas_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('entrada_id')->constrained('entradas')->onDelete('cascade');
             $table->foreignId('producto_id')->constrained('productos')->onDelete('cascade');
-            $table->decimal('precio_compra',12,3);
+            $table->bigInteger('precio_compra');
             $table->integer('cantidad');
             $table->timestamps();
         });
